+++ src/_pytask/console.py
@@ -45,7 +45,6 @@ __all__ = [
     "format_strings_as_flat_tree",
     "format_task_name",
     "get_file",
-    "is_jupyter",
     "render_to_string",
     "unify_styles",
 ]
@@ -200,7 +199,7 @@ def create_url_style_for_path(path: Path, edtior_url_scheme: str) -> Style:
     )
 
 
-def get_file(
+def get_file(  # noqa: PLR0911
     function: Callable[..., Any], skipped_paths: list[Path] | None = None
 ) -> Path | None:
     """Get path to module where the function is defined.
@@ -208,6 +207,11 @@ def get_file(
     When the ``pdb`` or ``trace`` mode is activated, every task function is wrapped with
     a decorator which we need to skip to get to the underlying task function. Thus, the
     special case.
+
+    Raises
+    ------
+    TypeError
+        If the object is a builtin module, class, or function.
 
     """
     if skipped_paths is None:
@@ -221,7 +225,19 @@ def get_file(
             return get_file(function.__wrapped__)
     source_file = inspect.getsourcefile(function)
     if source_file:
-        return Path(source_file)
+        # Handle functions defined in the REPL.
+        if "<stdin>" in source_file:
+            return None
+        # Handle lambda functions.
+        if "<string>" in source_file:
+            try:
+                return Path(function.__globals__["__file__"]).absolute().resolve()
+            except KeyError:
+                return None
+        # Handle functions defined in Jupyter notebooks.
+        if "ipykernel" in source_file or "ipython-input" in source_file:
+            return None
+        return Path(source_file).absolute().resolve()
     return None
 
 
@@ -287,26 +303,3 @@ def create_summary_panel(
         if counts[outcome_enum.FAIL]
         else outcome_enum.SUCCESS.style,
     )
-
-
-def is_jupyter() -> bool:  # pragma: no cover
-    """Check if we're running in a Jupyter notebook.
-
-    Copied from rich.
-
-    """
-    try:
-        get_ipython  # type: ignore[name-defined]  # noqa: B018
-    except NameError:
-        return False
-    ipython = get_ipython()  # type: ignore[name-defined]  # noqa: F821
-    shell = ipython.__class__.__name__
-    if (
-        "google.colab" in str(ipython.__class__)
-        or os.getenv("DATABRICKS_RUNTIME_VERSION")
-        or shell == "ZMQInteractiveShell"
-    ):
-        return True  # Jupyter notebook or qtconsole
-    if shell == "TerminalInteractiveShell":
-        return False  # Terminal running IPython
-    return False  # Other type (?)