+++ docs/source/changes.md
@@ -14,6 +14,7 @@ releases are available on [PyPI](https://pypi.org/project/pytask) and
 - {pull}`519` raises an error when builtin functions are wrapped with
   {func}`~pytask.task`. Closes {issue}`512`.
 - {pull}`522` improves the issue templates.
+- {pull}`523` refactors `_pytask.console._get_file`.
 
 ## 0.4.4 - 2023-12-04